# Ignite_V2_API_MVS_DB/openapi/ignite_openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ignite AI Agentic API (DB-backed)",
    "version": "2.0.0",
    "description": "Base URL: https://app.ingnityourgrowth.com\nAuth: header x-api-key"
  },
  "servers": [
    {
      "url": "https://app.ingnityourgrowth.com",
      "description": "Production"
    },
    {
      "url": "http://127.0.0.1:8000",
      "description": "Local Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    }
  },
  "paths": {
    "/api/data": {
      "get": {
        "summary": "List normalized data",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/valuescores": {
      "get": {
        "summary": "List value scores",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/insights": {
      "get": {
        "summary": "List insights",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/recommendations": {
      "get": {
        "summary": "List recommendations",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/segments": {
      "get": {
        "summary": "Get segments JSON",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/internal/ingest/data": {
      "post": {
        "summary": "Ingest data",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/internal/ingest/valuescores": {
      "post": {
        "summary": "Ingest value scores",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/internal/ingest/insights": {
      "post": {
        "summary": "Ingest insights",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/internal/ingest/recommendations": {
      "post": {
        "summary": "Ingest recommendations",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/internal/ingest/segments": {
      "post": {
        "summary": "Replace segments JSON",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  }
}